chore(sdk): merge v2 dsl code back to v1. (#5227)
* Merge v2 dsl back to v1 dsl

* fix tests

* fix tests and always attach container_spec to container_op

* address review comments

* fix a bad indentation
