+++ sdk/python/kfp/dsl/serialization_utils_test.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Tests for kfp.dsl.v2.serialization_utils module."""
+"""Tests for kfp.dsl.serialization_utils module."""
 import unittest
 
 from kfp.dsl import serialization_utils
